Move around methods and standardize dependency chain of datetime->timedelta.
